blogApp/pyproject.toml
[tool.poetry]
name = "blog-api"
version = "1.0.0"
description = "A FastAPI blog application with PostgreSQL, authentication, and likes/comments"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
PyJWT = "^2.8.0"
pwdlib = "^4.0.0"
alembic = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"